requirements.txt
Flask==1.0.2
marshmallow==2.18.1
PyJWT==1.7.0
PyMySQL==0.9.3
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.18
pytest==4.3.0
coverage==4.5.2
